+++ src/main/java/net/jasper/mod/PlayerAutomaClient.java
@@ -4,6 +4,7 @@ import net.fabricmc.api.ClientModInitializer;
 
 import net.jasper.mod.automation.*;
 import net.jasper.mod.gui.PlayerAutomaHUD;
+import net.jasper.mod.gui.RecordingSelectorScreen;
 import net.jasper.mod.util.keybinds.PlayerAutomaKeyBinds;
 import net.minecraft.client.MinecraftClient;
 import org.slf4j.Logger;
@@ -20,16 +21,13 @@ public class PlayerAutomaClient implements ClientModInitializer {
 	public static final String MOD_ID = "playerautoma";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID + "::client");
 	public static final String RECORDING_FOLDER_NAME = "recordings";
-	public static final String RECORDING_PREVIEW_FOLDER_NAME = "recording_previews";
 
 	public static final String PLAYERAUTOMA_FOLDER_PATH = Path.of(MinecraftClient.getInstance().runDirectory.getAbsolutePath(), MOD_ID).toString();
 	public static final String PLAYERAUTOMA_RECORDING_PATH = Path.of(PLAYERAUTOMA_FOLDER_PATH, RECORDING_FOLDER_NAME).toString();
-	public static final String PLAYERAUTOMA_RECORDING_PREVIEW_PATH = Path.of(PLAYERAUTOMA_FOLDER_PATH, RECORDING_PREVIEW_FOLDER_NAME).toString();
 
 	public static final String[] REQUIRED_FOLDERS = {
 			PLAYERAUTOMA_FOLDER_PATH,
 			PLAYERAUTOMA_RECORDING_PATH,
-			PLAYERAUTOMA_RECORDING_PREVIEW_PATH
 	};
 
 
@@ -69,6 +67,9 @@ public class PlayerAutomaClient implements ClientModInitializer {
 
 		// Register MenuPrevention. That allows to run Recording with minecraft in the background.
 		MenuPrevention.register();
+
+		// Load thumbnails from stored recordings to prevent lag
+		RecordingSelectorScreen.loadThumbnails();
 
 	}
 }